fix(ci): split identity service quality gate
